various integer overflow issues
git-svn-id: http://svn.vuze.com/public/client@29002 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/diskmanager/file/impl/FMFileAccessPieceReorderer.java
@@ -216,7 +216,7 @@ FMFileAccessPieceReorderer
 			
 					if ( physical_length > current_length ){
 					
-						long max_length = first_piece_length + (num_pieces-2)*piece_size + last_piece_length;
+						long max_length = first_piece_length + (num_pieces-2)*(long)piece_size + last_piece_length;
 					
 						physical_length = Math.min( physical_length, max_length );
